Manager refactoring
